requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.5.1
python-docx==0.8.11
requests==2.27.1
urllib3==1.26.9
